+++ etec-viagem/src/main/java/br/com/etechoracio/etec_viagem/repository/ViagemRepository.java
@@ -6,7 +6,7 @@ import br.com.etechoracio.etec_viagem.entity.Viagem;
 
 
 
-public interface ViagemRepository extends JpaRepository<Viagem, Integer> {
+public interface ViagemRepository extends JpaRepository<Viagem, Long> {
 
 	
 }
